+++ crates/symbolicator-service/src/services/caches/versions.rs
@@ -130,3 +130,11 @@ pub const SOURCEFILES_CACHE_VERSIONS: CacheVersions = CacheVersions {
     current: 1,
     fallbacks: &[],
 };
+
+/// Bundle Index Cache, with the following versions:
+///
+/// - `1`: Initial version.
+pub const BUNDLE_INDEX_CACHE_VERSIONS: CacheVersions = CacheVersions {
+    current: 1,
+    fallbacks: &[],
+};